requirements.txt
python-telegram-bot==20.8
requests==2.31.0
aiohttp==3.9.1
web3==6.11.3
apscheduler==3.10.4
sqlite3
asyncio
logging
python-dotenv==1.0.0
beautifulsoup4==4.12.2